issue/#56-stylish-ui * move existing components to their own directories * update imports
